+++ ats-mini/Common.h
@@ -14,7 +14,7 @@
 #define AUTHORS_LINE3  "Goshante, G8PTN (Dave), R9UCL (Max Arnold),"
 #define AUTHORS_LINE4  "Marat Fayzullin"
 #define APP_VERSION    226  // FIRMWARE VERSION
-#define EEPROM_VERSION 74   // EEPROM VERSION (forces reset)
+#define EEPROM_VERSION 75   // EEPROM VERSION (forces reset)
 
 // Modes
 #define FM            0
@@ -121,6 +121,15 @@ typedef struct
   const char *name;     // Color name
 } ClockColor;
 
+typedef struct
+{
+  uint8_t hour;
+  uint8_t minute;
+  uint8_t volume;
+  bool enabled;
+  bool triggered;
+} Alarm;
+
 //
 // Global Variables
 //
@@ -146,6 +155,7 @@ extern uint8_t  clockBrt;
 extern uint8_t  clockColorIdx;
 extern uint16_t currentSleep;
 extern uint8_t sleepModeIdx;
+extern Alarm alarms[2];
 extern bool zoomMenu;
 extern bool showTemperature;
 extern int8_t scrollDirection;