+++ ats-mini/Menu.h
@@ -44,7 +44,10 @@
 #define CMD_SLEEPMODE 0x2C00 // |
 #define CMD_LOADEIBI  0x2D00 // |
 #define CMD_WIFIMODE  0x2E00 // |
-#define CMD_ABOUT     0x2F00 //-+
+#define CMD_ALARM1    0x2F00 // |
+#define CMD_ALARM2    0x3000 // |
+#define CMD_ALARMVOL  0x3100 // |
+#define CMD_ABOUT     0x3200 //-+
 
 // UI Layouts
 #define UI_DEFAULT  0
@@ -132,5 +135,7 @@ void doCal(int dir);
 void doStep(int dir);
 void doMode(int dir);
 void doBand(int dir);
+void doAlarmTime(int idx, int dir);
+void doAlarmVol(int dir);
 
 #endif // MENU_H